Stuff is working

+++ mob.cpp
@@ -1,6 +1,6 @@
 #include "mob.h"
 
-Mob::Mob(std::string file, int locationX, int locationY, int sizeX, int sizeY, b2World &world)
+Mob::Mob(std::string file, float locationX, float locationY, float sizeX, float sizeY, b2World &world)
     : position(locationX, locationY),
       size(sizeX,sizeY)
 {
@@ -19,6 +19,8 @@ void Mob::Update(windowTransform transform)
 {
     // Updates SFML sprite with b2Body position and rotation
     sprite.setPosition(transform.transformX(body->GetPosition().x), transform.transformY(body->GetPosition().y));
+    //printf("%f, %d; %f, %d\n",body->GetPosition().x, transform.transformX(body->GetPosition().x)
+    //                 ,body->GetPosition().y, transform.transformY(body->GetPosition().y));
     sprite.setRotation(transform.transformAngle(body->GetAngle()));
     auto bounds = sprite.getLocalBounds();
     sprite.setScale(transform.transformWidth(size.x) / bounds.width, transform.transformHeight(size.y) / bounds.height);
